requirements.txt
Flask
flask-jwt-extended
requests
rsa
pycryptodome
pycryptodomex